requirements.txt
Flask ~= 1.1.2
sqlalchemy ~= 1.3.20
psycopg2 ~= 2.8.6
uuid ~= 1.30

pyyaml ~= 5.3.1
